MLService/requirements.txt
falcon
boto3
torch
numpy
Pillow
transformers
moviepy
loralib
glob
pickle
openai
openai